+++ test/test_scene_stack.py
@@ -1,10 +1,15 @@
 import numpy as num
 import logging
 import time
+
 from kite.scene import TestScene
 from kite.scene_stack import TSScene, SceneStack
+from kite.quadtree import QuadNode
+
+from kite.spool import spool
 
 logging.basicConfig(level=logging.DEBUG)
+QuadNode.MIN_PIXEL_LENGTH_NODE = 32
 
 t0 = time.time()
 dt = 60*60*24*365.25 / 2
@@ -16,10 +21,28 @@ class TSTestScene(TestScene, TSScene):
     def createSine(cls, *args, **kwargs):
         return super().createSine(*args, **kwargs)
 
+    @classmethod
+    def createRandom(cls, *args, **kwargs):
+        return super().createRandom(*args, **kwargs)
+
+    @classmethod
+    def createGauss(cls, *args, **kwargs):
+        return super().createGauss(*args, **kwargs)
+
+    @classmethod
+    def createFractal(cls, *args, **kwargs):
+        return super().createFractal(*args, **kwargs)
+
 
 nscenes = 10
-scenes = {t0 - its*dt: TSTestScene.createSine()
-          for its in range(nscenes)}
+scenes = {}
+
+for isc in range(nscenes):
+    t = t0 - isc * dt
+    sc = TSTestScene.createRandom(256, 256)
+    sc.addNoise(1.)
+
+    scenes[t] = sc
 
 
 stack = SceneStack()
@@ -38,3 +61,6 @@ def test_set_scene_to():
     times = num.linspace(tmin, tmax, 30)
     for ts in times:
         stack.set_scene_to(ts)
+
+
+spool(stack)