+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to Kite will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+- Import of SNAP products
+
+### Fixed
+- Spool: improved UI
+
 ## [1.3.0] 3. December 2019
 
 ### Added